Move create and update draft classes
Moved out of commands since they are not really commands and moved them
into models so that they are visible for reuse.
